requirements.txt
streamlit
pandas
numpy
plotly
psycopg2-binary
bcrypt

